+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - proguard: Added a mandatory `index` field to `JvmFrame` (#1411) by @loewenheim
 - proguard: Support for source contex (#1414) by @loewenheim
 - proguard: Field `lineno` on `JvmFrame` is now optional (#1417) by @loewenheim
-- proguard: add support for frame signature (#1425 )by @viglia
+- proguard: add support for frame signature (#1425 ) by @viglia
+- proguard: add the translated and deobfuscated signature even in the case that the the whole frame could not be remapped (#1427) by @viglia
 
 ### Dependencies